requirements.txt
Django>=4.2
djangorestframework
requests
geopy
python-dotenv